changed all references of cleangym to heatenginegym
