+++ heatenginegym/__init__.py
@@ -2,55 +2,55 @@ from gym.envs.registration import register
 
 register(
     id='Carnot-v0',
-    entry_point='cleangym.carnot_lim:CarnotEnv',
+    entry_point='heatenginegym.carnot_lim:CarnotEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Carnot-v1',
-    entry_point='cleangym.carnot:CarnotEnv',
+    entry_point='heatenginegym.carnot:CarnotEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Carnot-v2',
-    entry_point='cleangym.carnot_dV:CarnotEnv',
+    entry_point='heatenginegym.carnot_dV:CarnotEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Beeler-v0',
-    entry_point='cleangym.beeler:BeelerEnv',
+    entry_point='heatenginegym.beeler:BeelerEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Beeler-v1',
-    entry_point='cleangym.beeler_dV:BeelerEnv',
+    entry_point='heatenginegym.beeler_dV:BeelerEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Stirling-v0',
-    entry_point='cleangym.stirling:StirlingEnv',
+    entry_point='heatenginegym.stirling:StirlingEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Stirling-v1',
-    entry_point='cleangym.stirling_dV:StirlingEnv',
+    entry_point='heatenginegym.stirling_dV:StirlingEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Otto-v0',
-    entry_point='cleangym.otto:OttoEnv',
+    entry_point='heatenginegym.otto:OttoEnv',
     max_episode_steps=500,
 )
 
 register(
     id='Otto-v1',
-    entry_point='cleangym.otto_dV:OttoEnv',
+    entry_point='heatenginegym.otto_dV:OttoEnv',
     max_episode_steps=500,
 )
